Combine implementations & use inside of tests
