+++ relapse/rest/client/openid.py
@@ -99,4 +99,5 @@ class IdTokenServlet(RestServlet):
 
 
 def register_servlets(hs: "HomeServer", http_server: HttpServer) -> None:
-    IdTokenServlet(hs).register(http_server)
+    if not hs.config.experimental.msc3861.enabled:
+        IdTokenServlet(hs).register(http_server)
