+++ tests/replication/_base.py
@@ -349,12 +349,15 @@ class BaseMultiWorkerStreamTestCase(unittest.HomeserverTestCase):
         store.db_pool._db_pool = self.database_pool._db_pool
 
         # Set up a resource for the worker
-        resource = self.create_test_resource(worker_hs)
-
+        listener_config = worker_hs.config.server.listeners[0]
+        assert listener_config.http_options is not None
+        resource = worker_hs.resource_for_listener(
+            listener_config.http_options.resources
+        )
         self._hs_to_site[worker_hs] = RelapseSite(
             logger_name="relapse.access.http.fake",
             site_tag=f"{worker_hs.config.server.server_name}-{worker_hs.get_instance_name()}",
-            config=worker_hs.config.server.listeners[0],
+            config=listener_config,
             resource=resource,
             server_version_string="1",
             max_request_body_size=8192,
